requirements.txt
openai>=1.14.0
pandas>=2.0.0
pyarrow>=14.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
